Room scaffolding
To get started with Room, add a dummy class Foo and a database to
persist it.

Do not persist the database scheme. Before the database is somewhat stable,
there is no point in versioning the database. However, when this changes,
modify ChronosDatabase's 'exportSchema' and add the following to
app/build.gradle:

android {
    defaultConfig {
        ...
        javaCompileOptions {
            annotationProcessorOptions {
                arguments = ["room.schemaLocation": "$projectDir/schemas".toString()]
            }
        }
    ...
}

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>