Dependencies: add ThreeTen Android Backport (1.0.5)
This project will need timestamps with timezone data. The standard
java.time classes are a mess before Java 8 which introduces
OffsetDateTime. Because Java 8 is not available until SDK 26, and this
project is currently on min SDK 24, OffsetDateTime and friends are not
available out of the box.

Jake Wharton's ThreeTen Android Backport (threetenabp) project is a
backport of Java 8's java.time classes, with Android in mind. Add a
dependency to threetenabp.

When this project bumps its min SDK to 26, the threetenabp classes
should be replaced by the standard Java classes. This is done by:

  1. Removing the call to AndroidThreeTen.init (in
     ChronosApplication.kt)

  2.  s/import org.threeten.bp/import java.time/

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>