Connect recycler view to database via LiveData
Add a middle layer to connect the higher layer (the recycler view) with
the lower layer (the database). The UI now works exclusively with
LiveData objects. ViewModels maintains these objects while UI components
are recreated as part of the activity lifecycle.

Introduce a middle layer class Repository to sit between ViewModels and
the database. In the current incarnation, with only a single data
source, the Repository class is just another proxy layer, but in the
future it will be used to implement the repository pattern.

Disclaimer and credit: the design was heavily influenced by the
BasicSample in the Android Architecture Components samples project [1].
Some code was directly copied from [1], both top level functions:

  - getValue in ChronosDatabaseTest.kt
  - ioThread in Repository.kt

1.  https://github.com/googlesamples/android-architecture-components.git

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>