Add navigation drawer
Add a navigation drawer. When the user clicks a menu item, the main
activity switches the currently displayed fragment.

In this commit, the logic to switch fragments is triggered directly from
the click listener. A future change will improve the application's
design by introducing a lower layer Navigator class: when the user
clicks a menu item, a request will be sent to the Navigator which then
decides what fragment the main activity should switch to.

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>