Prevent irrelevant LiveData emissions from reaching the UI
A LiveData returned from a DAO will emit its value on any database
change, even if the change did not affect the values observed by the
LiveData.

Add an intermediate MediatorLiveData to remember the latest emitted data
and to prevent further emissions until there is a change in the data.

Since this trick works best with a method in the DAO, change FooDao from
an interface to an abstract class (Room will accept either).

This change is based on an article by Florina Muntenescu [1]; the code
is copied from the article (and modified slightly to reduce the
complexity of the if expression but with the downside that 'initialized'
is assigned on each emission).

1. https://medium.com/google-developers/7-pro-tips-for-room-fbadea4bfbd1#5e38

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>