Introduce Block and related classes
Add a new interface 'Block' to represent a block of time spent on a
given task, such as working for a client or traveling. Implement the
interface in a new class 'BlockEntity'.

Classes that end in Entity (like class BlockEntity) belong to the
database layer. The rest of the application works on a more generic type
(like interface Block).

Thanks to the way Kotlin generics work, there is no need to create new
objects as BlockEntities bubble up through the upper layers via the
BlockViewModel: a straight-forward cast will do:

    list: List<BlockEntity> -> list as List<Block>

Signed-off-by: Mårten Kongstad <marten.kongstad@gmail.com>